requirements.txt
aiogram==3.11.0
flask==3.0.3
sqlalchemy==2.0.29
python-dotenv==1.0.1
aiohttp==3.9.3